requirements.txt
# google-cloud==0.34.0
google-cloud-bigquery==1.24.0
google-cloud-storage==1.28.1
google-api-python-client==1.8.3
google-oauth==1.0.1
PyYAML==5.3.1